app/requirements.txt
sqlalchemy
python-dotenv
flask
flask-sqlalchemy
passlib
pymysql
mysql-connector
numpy 
xgboost
psycopg2
